requirements.txt
Package             Version
------------------- -------
asgiref             3.10.0
dataclasses         0.6
Django              5.2.7
djangorestframework 3.16.1
dnspython           2.8.0
pip                 24.0
pymongo             4.15.3
pytz                2025.2
setuptools          65.5.0
sqlparse            0.5.3
tzdata              2025.2